Pypi expects a __main__.py file in the primary module. Create backwards compat in parent directory.
